// src/types/Commands.ts
export enum Commands {
  Mouse_up = 'mouse_up',
  Mouse_right = 'mouse_right',
  Mouse_down = 'mouse_down',
  Mouse_left = 'mouse_left',
  Mouse_position = 'mouse_position',
  Draw_circle = 'draw_circle',
  Draw_square = 'draw_square',
  Draw_rectangle = 'draw_rectangle',
  Prnt_scrn = 'prnt_scrn',
}